Adds login fixture and uses it
logged_in_user is a fixture which creates a user and monkeypatches/mocks
the current_user attribute in flask-login and views to make it appear as if
the test client is logged in. This is required to test the socketio
code for lecturers, as it requres you to be logged in.
Setting LOGIN_DISABLED=True (Flask-Login config value) and checking for
 it in the lecturer_recv view could be another way to test
without needing to log in.