Changes the splash page to a search for courses
Students can search for courses, and the search result
is updated in real time. When the search field is empty,
all courses are hidden.

Some refactoring is done on the functions which retrieves
objects from the database, to support searching on non-PK
attributes. This refactoring also adds a  more explicit and
robust way of specifying parameters for the filtering and
instantiation of objects (using keyword arguments).