Add backend-part of predefined responses
Keywords are extracted by an external API.
After the keyword has been extracted, if is searched for a matching
response in the DB. The response is sent as the third element in
the question response socketIO event (null if none was found).
The matching response is also saved together with the question.
There is also a listener for socketIO for adding new keyword-responses
The extra flags in requirements.txt are to be able to
build pillow without extra dependecies.